Split out Serversettings to its own file
