+++ modules/express-engine/schematics/install/files/server-tools/server-settings.interface.ts
@@ -1,0 +1,14 @@
+
+
+export interface ServerSettings {
+  /** is SSL going to be used? */
+  useSSL: boolean;
+  /** Full path to private key file. */
+  sslKey: string;
+  /** Full path to public cert file. */
+  sslCert: string;
+  /** The hostname to bind on. */
+  hostName: string;
+  /** The port number the server will be served on. */
+  port: number;
+}
